refactored examples and fixed compilation issue

+++ src/game/Bomberman.cpp
@@ -1,5 +1,7 @@
 #include "Bomberman.hpp"
 
+#include <gsl/gsl_util>
+
 Bomberman::Bomberman(physics::PhysicsId phys_id, graphics::renderer_id rend_id,
                      std::unique_ptr<HumanPlayer> player,
                      std::shared_ptr<physics::PhysicsEngine> physics,
@@ -13,7 +15,7 @@ Bomberman::Bomberman(physics::PhysicsId phys_id, graphics::renderer_id rend_id,
 
 void Bomberman::update(DeltaTime dt)
 {
-    constexpr float speedup_factor{100};
+    constexpr double speedup_factor{100};
     auto position = physics_engine->get_position(physics_id);
     renderer_pool->set_position(renderer_id, position);
 
@@ -23,8 +25,10 @@ void Bomberman::update(DeltaTime dt)
     }
 
     const auto new_direction = input->get_direction();
-    position.x += (speedup_factor * new_direction.x) * dt.count();
-    position.y += (speedup_factor * new_direction.y) * dt.count();
+    position.x +=
+        gsl::narrow_cast<float>(speedup_factor * new_direction.x * dt.count());
+    position.y +=
+        gsl::narrow_cast<float>(speedup_factor * new_direction.y * dt.count());
 
     physics_engine->set_position(physics_id, position);
 }
